refactor: 모든 Atom 컴포넌트 별칭 경로 수정 (#78)
