+++ src/shared/ui/atoms/floating-button/variant/IconFloatingButton.tsx
@@ -1,6 +1,6 @@
 import { type PressableProps } from "react-native";
-import Button from "../../button";
-import { IconName } from "../../icon/variants";
+import { Button } from "@shared/ui/atoms";
+import type { IconName } from "@shared/ui/atoms/icon/variants";
 import { preset } from "../foundation/preset";
 
 export interface Props extends PressableProps {
