+++ src/shared/ui/atoms/tag/index.tsx
@@ -1,6 +1,6 @@
 import { View, ViewProps } from "react-native";
-import { ParkingStatus } from "@/shared/types/parking-status";
-import Text from "@shared/ui/atoms/text";
+import { ParkingStatus } from "@shared/types/parking-status";
+import { Text } from "@shared/ui/atoms";
 import { TagVariant } from "./variants";
 
 // Tag 컴포넌트에 대한 Props 타입 정의
