+++ src/shared/ui/atoms/text/index.tsx
@@ -1,6 +1,6 @@
 import clsx from "clsx";
 import { Text as RNText, TextProps } from "react-native";
-import { TypographyKey, typographyTokens } from "@shared/tokens/typography";
+import { type TypographyKey, typographyTokens } from "@shared/tokens/typography";
 
 interface Props extends TextProps {
   typography?: TypographyKey;
